requirements.txt
# Core AI
anthropic>=0.40.0

# Discord
discord.py>=2.3.0

# Todoist
httpx>=0.27.0

# Calendar (iCloud CalDAV)
caldav>=1.3.9
icalendar>=5.0.13

# Memory
chromadb>=0.5.0

# Scheduler
APScheduler>=3.10.0

# Config
pydantic-settings>=2.0.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
